packages/python/agent/pyproject.toml
[project]
name = "omni-agent"
version = "0.5.0-dev"
description = "Omni Agent - Thin MCP/Web Client Interface"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # Core + Foundation (via transitive deps)
    "omni-core",
    "omni-mcp>=0.1.0",
    # Web & CLI
    "starlette>=0.40.0",
    "uvicorn>=0.30.0",
    "uvloop>=0.22.0; sys_platform != 'win32'",
    "httptools>=0.7.1",
    "typer>=0.15.0",
    "aiohttp>=3.11.0",
    # Logging (pervasive)
    "structlog>=25.5.0",
    # Stress testing
    "psutil>=5.9.0",
    "pyarrow>=23.0.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/omni"]


[project.scripts]
omni = "omni.agent.cli.app:entry_point"


[tool.uv.sources]
omni-core = { workspace = true }
omni-mcp = { workspace = true }


[dependency-groups]
dev = [
    "omni-test-kit",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "pytest-xdist>=3.6.0",
    "pytest-randomly>=4.0.0",
    "pytest-benchmark>=4.0.0",
    "dirty-equals>=0.11.0",
    "polyfactory>=3.2.0",
    "ruff>=0.14.0",
]


# --- Testing Configuration ---
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-n", "auto",
    "--tb", "short",
    "-q",
    "-p", "no:benchmark",
]
pythonpath = ["src"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:Benchmarks are automatically disabled:pytest_benchmark.logger.PytestBenchmarkWarning",
    "ignore:Module already imported so cannot be rewritten; omni\\.test_kit\\.fixtures:pytest.PytestAssertRewriteWarning",
]
